requirements.txt
tiktoken
transformers
torch
wikipedia-api
deep-translator
nltk
matplotlib
seaborn
python-dotenv
